+++ app/src/main/java/io/github/domi04151309/home/api/HueAPI.kt
@@ -1,4 +1,4 @@
-package io.github.domi04151309.home.helpers
+package io.github.domi04151309.home.api
 
 import android.content.Context
 import android.content.Intent
@@ -18,6 +18,7 @@ import io.github.domi04151309.home.data.RequestCallbackObject
 import io.github.domi04151309.home.custom.CustomJsonArrayRequest
 import io.github.domi04151309.home.data.ListViewItem
 import io.github.domi04151309.home.data.UnifiedRequestCallback
+import io.github.domi04151309.home.helpers.Global
 import io.github.domi04151309.home.interfaces.HomeRecyclerViewHelperInterface
 
 class HueAPI(
