+++ app/src/main/java/io/github/domi04151309/home/api/ShellyAPI.kt
@@ -1,4 +1,4 @@
-package io.github.domi04151309.home.helpers
+package io.github.domi04151309.home.api
 
 import android.content.Context
 import android.util.Log
@@ -7,6 +7,8 @@ import com.android.volley.Response
 import com.android.volley.toolbox.JsonObjectRequest
 import io.github.domi04151309.home.custom.JsonObjectRequestAuth
 import io.github.domi04151309.home.data.UnifiedRequestCallback
+import io.github.domi04151309.home.helpers.DeviceSecrets
+import io.github.domi04151309.home.helpers.Global
 import io.github.domi04151309.home.interfaces.HomeRecyclerViewHelperInterface
 
 class ShellyAPI(
@@ -36,12 +38,16 @@ class ShellyAPI(
                             )
                         },
                         { error ->
-                            callback.onItemsLoaded(UnifiedRequestCallback(null, deviceId, Global.volleyError(c, error)), null)
+                            callback.onItemsLoaded(UnifiedRequestCallback(null, deviceId,
+                                Global.volleyError(c, error)
+                            ), null)
                         }
                     ))
                 },
                 { error ->
-                    callback.onItemsLoaded(UnifiedRequestCallback(null, deviceId, Global.volleyError(c, error)), null)
+                    callback.onItemsLoaded(UnifiedRequestCallback(null, deviceId,
+                        Global.volleyError(c, error)
+                    ), null)
                 }
             )
             2 -> JsonObjectRequest(
@@ -59,12 +65,16 @@ class ShellyAPI(
                             )
                         },
                         { error ->
-                            callback.onItemsLoaded(UnifiedRequestCallback(null, deviceId, Global.volleyError(c, error)), null)
+                            callback.onItemsLoaded(UnifiedRequestCallback(null, deviceId,
+                                Global.volleyError(c, error)
+                            ), null)
                         }
                     ))
                 },
                 { error ->
-                    callback.onItemsLoaded(UnifiedRequestCallback(null, deviceId, Global.volleyError(c, error)), null)
+                    callback.onItemsLoaded(UnifiedRequestCallback(null, deviceId,
+                        Global.volleyError(c, error)
+                    ), null)
                 }
             )
             else -> null
