+++ app/src/main/java/io/github/domi04151309/home/api/SimpleHomeAPI.kt
@@ -1,4 +1,4 @@
-package io.github.domi04151309.home.helpers
+package io.github.domi04151309.home.api
 
 import android.content.Context
 import android.util.Log
@@ -8,6 +8,9 @@ import io.github.domi04151309.home.helpers.Global.volleyError
 import io.github.domi04151309.home.R
 import io.github.domi04151309.home.data.ListViewItem
 import io.github.domi04151309.home.data.UnifiedRequestCallback
+import io.github.domi04151309.home.helpers.Commands
+import io.github.domi04151309.home.helpers.Devices
+import io.github.domi04151309.home.helpers.Global
 import io.github.domi04151309.home.interfaces.HomeRecyclerViewHelperInterface
 import org.json.JSONException
