+++ app/src/main/java/io/github/domi04151309/home/api/UnifiedAPI.kt
@@ -1,9 +1,10 @@
-package io.github.domi04151309.home.helpers
+package io.github.domi04151309.home.api
 
 import android.content.Context
 import com.android.volley.RequestQueue
 import com.android.volley.toolbox.Volley
 import io.github.domi04151309.home.data.UnifiedRequestCallback
+import io.github.domi04151309.home.helpers.Devices
 import io.github.domi04151309.home.interfaces.HomeRecyclerViewHelperInterface
 
 open class UnifiedAPI(
