+++ app/src/test/java/io/github/domi04151309/home/helpers/EspEasyAPIParserTest.kt
@@ -2,6 +2,7 @@ package io.github.domi04151309.home.helpers
 
 import android.content.res.Resources
 import io.github.domi04151309.home.R
+import io.github.domi04151309.home.api.EspEasyAPIParser
 import org.junit.Assert
 import org.json.JSONObject
 import org.junit.Test
